Add version change comments on PRs with `whippet.lock` changes
This commit expands the workflow so that it runs `whippet deps describe`
on the head and target branch for any PR that includes changes to
`whippet.lock`. The output of the two `whippet deps describe` is then
passed into `diff`, and the output posted in a comment on the PR.

The result is that any PRs that amend `whippet.lock` should receive a
comment describing what version changes will be made as a result of the
PR (something that isn't clear at the moment, as `whippet.lock` just
refers to commit hashes rather than version tags). This functionality
could be expanded in future so that e.g. PRs that only modify
`whippet.lock` and don't introduce any major version changes are
auto-merged.

There's a few things to note about the way this will work:

- The workflow will need read access to not only the repo containing the
  PR, but also all the Whippet dependencies the `whippet.lock` file
  references. The default GitHub token used in a workflow only has
  permission to read the source repo, so we have to pass in a GitHub
  token with at least read permission for the source repo and all our
  `dxw-wordpress-plugin` repos.
- `diff` produces an exit code of 1 if it finds differences, which by
  default causes the workflow to fail, so we run `set +e` before running
  `diff` the `set -e` after.
- If a PR is subsequently updated with more changes to `whippet.lock`,
  the original comment will be updated. We could modify this so the
  original comment is removed and replaced with a new comment if we want
  to.
- The workflow uses a couple of third-party actions
  (`dorny/paths-filter@v2` and
  `marocchino/sticky-pull-request-comment@v2`) that we haven't used
  before, but both seem well-maintained and widely used.