Use default token for checking out source repo
Before: we were using the GitHub token passed as a secret to checkout
both the calling repo and any Whippet dependency repos. This didn't work
when using a fine-grained PAT, as they only work over https and the
GitHub API (i.e. not when checking out the calling repo).

Now: we use the default token generated for the workflow run to checkout
the calling repo (this token only has access to the calling repo), and
the passed secret as the token when authenticating with GitHub CLI for
checking out other repos. This gives us finer control over permissions,
as it allows us to use fine-grained tokens, with read-only access.