Enforce references for whippet dependencies
As part of the work to automate plugin updates, we want to ensure that
we don't accidentally deploy a breaking change when running `whippet
deps update`. The best way to do this currently is to pin all plugins to
a specific git reference (normally the major version tag for the current
release).

This change ensures that the workflow will fail if a file in
`whippet.json` does not have a reference specified.

See https://github.com/dxw/whippet/pull/245 for where this feature was
added to Whippet.